Prepare the sequences with repeat elements

The function will extract the genomic annotations from TxDb object, extract the sequences from BSgenome object, and generate the annotation for repeat elements.

prepareSeq(genome, txdb, rmsk, peaks, minWidth = 31L, verbose = FALSE, ...)

Arguments

genome

A BSgenome object

txdb

A TxDb object

rmsk

The RepeatMasker GRanges object

peaks

An GRanges object to represent the peak list.

minWidth

Minimal width for the repeat elements in the intergenic annotation.

verbose

Print the message or not.

...

Not use.

Value

An DataFrame with seqname and their repeat element annotations

Examples

# library(AnnotationHub)
# ah <- AnnotationHub()
# (rmsk <- query(ah, c('RepeatMasker', 'Zebrafish')))
# rmsk <- rmsk[['AH98980']]
rmsk <- readRDS(system.file('extdata', 'danRer11_rmsk_chr25_sub.RDS',
                             package='ratte'))

library(BSgenome.Drerio.UCSC.danRer11) # for sequences
library(TxDb.Drerio.UCSC.danRer11.refGene) # for genomic annotation
prepareSeq(Drerio,
           TxDb.Drerio.UCSC.danRer11.refGene,
           rmsk)
#> Warning: some group names are NAs or duplicated
#> DataFrame with 8783 rows and 9 columns
#>                     seqname     feature      repName       repClass
#>                 <character> <character>  <character>    <character>
#> 1    dd7050e8ed56cd21c9a3..        exon       A-rich Low_complexity
#> 2    dd7050e8ed56cd21c9a3..        exon DNA-2-24B_DR            DNA
#> 3    3e7e12f683a34f764248..        exon       (GGT)n  Simple_repeat
#> 4    c3e49fbbadb65ef5746a..        exon       (CCA)n  Simple_repeat
#> 5    4e3f05db4fa02bcba322..        exon       (CAG)n  Simple_repeat
#> ...                     ...         ...          ...            ...
#> 8779 0c45ac71ec1bff01af37..  intergenic     Tc1N1_DR            DNA
#> 8780 3d0024bdcc1f065684ee..  intergenic     Tc1N1_DR            DNA
#> 8781 75ab38b962fed592ed30..  intergenic   ENSPM-6_DR            DNA
#> 8782 75ab38b962fed592ed30..  intergenic ERV1-N3-I_DR            LTR
#> 8783 7035a5634fcfd9c01b5b..  intergenic      ASAT_CY      Satellite
#>           repFamily         N                    idx      tx_name
#>         <character> <integer>            <character>  <character>
#> 1    Low_complexity         1 chr25:11005150-11005.. NM_001327884
#> 2               DNA         1 chr25:11006165-11006.. NM_001327884
#> 3     Simple_repeat         1 chr25:12689061-12689.. NM_001122702
#> 4     Simple_repeat         1 chr25:12849652-12849.. NM_001128278
#> 5     Simple_repeat         1 chr25:13406054-13406..    NM_200870
#> ...             ...       ...                    ...          ...
#> 8779     CMC-EnSpm?         1 chr25:14990246-14990..           NA
#> 8780     CMC-EnSpm?         1 chr25:14990583-14990..           NA
#> 8781      CMC-EnSpm         1 chr25:14993932-14994..           NA
#> 8782            LTR         1 chr25:14993948-14994..           NA
#> 8783      Satellite         1 chr25:14999170-14999..           NA
#>                          seq
#>               <DNAStringSet>
#> 1    AAAGAAGTGT...AGCTGTGCAT
#> 2    AAAGAAGTGT...AGCTGTGCAT
#> 3    ATGTCCACTG...TTTTATTTGA
#> 4    TCTGTGAGCC...TAATACTGTT
#> 5    CAGCAACACA...GATGAAATGA
#> ...                      ...
#> 8779 TGTTTATTTT...GACATTGTAG
#> 8780 TTTGAAATAA...ATTTTACAGT
#> 8781 CTCTCTCATT...TTTGGATGTG
#> 8782 CTCTCTCATT...TTTGGATGTG
#> 8783 TTAACATTTG...GTTTTAAATT